update the post so it in it's own modal
just need to change the input as users should be able to input a id for a post

+++ BackEnd/BackEndSD/src/main/java/com/example/supabase/models/Post.java
@@ -8,29 +8,51 @@ public class Post {
 
 	@Id
     private String id;
-    private Long userId; // Reference to the user stored in PostgreSQL
+    private int userId; // Reference to the user stored in PostgreSQL
     private String img; // URL or base64 string of the image
     private String description; // Message or caption for the post
 
-    // Constructors
+    
+    // No-argument constructor 
     public Post() {}
 
-    public Post(Long userId, String img, String description) {
+    // Constructor
+    public Post(int userId, String img, String description) {
         this.userId = userId;
         this.img = img;
         this.description = description;
     }
 
     // Getters & Setters
-    public String getId() { return id; }
-    public void setId(String id) { this.id = id; }
+    public String getId() { 
+    	return id; 
+    }
+    
+    public void setId(String id) { 
+    	this.id = id; 
+    }
 
-    public Long getUserId() { return userId; }
-    public void setUserId(Long userId) { this.userId = userId; }
+    public int getUserId() { 
+    	return userId; 
+    }
+    
+    public void setUserId(int userId) { 
+    	this.userId = userId; 
+    }
 
-    public String getImg() { return img; }
-    public void setImg(String img) { this.img = img; }
+    public String getImg() { 
+    	return img; 
+    }
+    
+    public void setImg(String img) { 
+    	this.img = img; 
+    }
 
-    public String getDescription() { return description; }
-    public void setDescription(String description) { this.description = description; }
+    public String getDescription() { 
+    	return description; 
+    }
+    
+    public void setDescription(String description) { 
+    	this.description = description; 
+    }
 }
